organized nanny components
Added all nanny components under one component
